# By bramas (7) and piacibel (4)
# Via bramas
* 'master' of git+ssh://scm.gforge.inria.fr//gitroot/scalfmm/scalfmm:
  Add a name to critical section (even if they should not be called at the same time but to ensure that they cannot interfere with some others)
  Add critical when creating kernel for each thread to ensure that the thread are in a protected environment
  minor fix on that test
  Add operation to the vector to ensure that object without default constructor can be user with it
  Both version (user defined kernel and interpolation kernel) provides same features
  oops missmatch in the deleter for the smart ptr
  Remove some warnings and add the patch for the omp to starpu auto version
  Remove some warnings and add the patch for the omp to starpu auto version
  remove some conversion warnings
  Forgot 1 modifocation
  Update in API in order to remove memory leaks (changes added to FChebInterpolator too.)

+++ Src/Containers/FMpiBufferWriter.hpp
@@ -37,7 +37,7 @@ class FMpiBufferWriter : public FAbstractBufferWriter {
     /** Test and exit if not enought space */
     void expandIfNeeded(const size_t requestedSpace) {
         if( arrayCapacity < int(currentIndex + requestedSpace) ){
-            arrayCapacity = int((currentIndex + requestedSpace + 1) * 1.5);
+            arrayCapacity = int(double(currentIndex + requestedSpace + 1) * 1.5);
             char* arrayTmp = new char[arrayCapacity];
             memcpy(arrayTmp, array.get(), sizeof(char)*currentIndex);
             array.reset(arrayTmp);
